Updated expression_pipeline/README.md
Updated the expression pipeline README to include recent changes to the scripts,
including the incorporation of getArgs() to parse command line arguments. I also
adjusted the examples a bit to make it more clear where the commands should be run,
which commands should be run on a compute node, and which can safely be run on the
head node.