Fixed issues that arose when a control was run on one plate and not the other
When a cell type was composed of 2+ plates and a specific control only run on
one (e.g., DMSO for DP11), the script was creating a dummy vector of 0 counts on
the opposite plate, resulting in 0 heterozygotes and an error during downstream
ASE analysis. These 0-vectors are now dropped if they appear. Additionally,
improved the code creating vectors of treatment names and treatment IDs (used in
plots and output filenames), which now accounts for this issue.