Fixed issue where malformed bedfile was causing an error
For some samples, the intersectBed stage of UnionExtractFields was failing. As
best I could determine, the issue has to do with the gzipped temp file created,
which looked fine on inspection, but the intersectBed command failed, stating
incorrect number of rows. I adjusted the code to not gzip the file in the first
place, which has resolved the issue.