Fixed bug in matching treatment and control for each SNP
This updated fixes an issue in the MESH input table. MESH requires control and
treatment data on the same line, and so the master table must be reshaped
accordingly. The previous code was incorrectly matching control and treatment,
resulting in NAs for many of the control samples, which ultimately affected
the overall MESH run.